+++ python/instrumentation/openinference-instrumentation-google-genai/examples/generate_content.py
@@ -3,7 +3,9 @@ import os
 
 from google import genai
 from google.genai.types import Content, GenerateContentConfig, Part
-from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
+from opentelemetry.exporter.otlp.proto.http.trace_exporter import (
+    OTLPSpanExporter,  # type: ignore[import-not-found]
+)
 from opentelemetry.sdk import trace as trace_sdk
 from opentelemetry.sdk.trace.export import ConsoleSpanExporter, SimpleSpanProcessor