+++ python/instrumentation/openinference-instrumentation-langchain/tests/test_image_in_message.py
@@ -34,7 +34,7 @@ def test_image_in_message(
     assert attributes.pop(SpanAttributes.INPUT_VALUE, None)
     assert attributes.pop(SpanAttributes.OUTPUT_MIME_TYPE, None)
     assert attributes.pop(SpanAttributes.OUTPUT_VALUE, None)
-    assert attributes.pop(SpanAttributes.LLM_MODEL_NAME, None)
+    assert attributes.pop(SpanAttributes.LLM_MODEL_NAME, None) == "gpt-4o"
     assert attributes.pop(SpanAttributes.LLM_INVOCATION_PARAMETERS, None)
     if LANGCHAIN_VERSION >= (0, 2):
         assert attributes.pop(SpanAttributes.METADATA, None)