+++ js/packages/openinference-vercel/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @arizeai/openinference-vercel
 
+## 2.2.2
+
+### Patch Changes
+
+- Updated dependencies [34a4159]
+  - @arizeai/openinference-semantic-conventions@2.1.0
+  - @arizeai/openinference-core@1.0.4
+
 ## 2.2.1
 
 ### Patch Changes